+++ Sandbox/src/Audio/ChannelMixer.h
@@ -6,7 +6,7 @@ class ChannelMixer
 public:
 	ChannelMixer(uint32_t channelTargetCount);
 
-	void MixChannels(int16_t** sampleDataPtr, size_t* sampleCountPtr, uint32_t* channelCountPtr);
+	void MixChannels(int16_t** sampleDataPtr, uint64_t* sampleCountPtr, uint32_t* channelCountPtr);
 
 protected:
 	uint32_t channelTargetCount;
