+++ Sandbox/src/Audio/DummySampleProvider.h
@@ -4,8 +4,8 @@
 class DummySampleProvider : public ISampleProvider
 {
 public:
-	virtual size_t ReadSamples(int16_t* bufferToFill, size_t sampleOffset, size_t samplesToRead) override;
-	virtual size_t GetSampleCount() override;
+	virtual int64_t ReadSamples(int16_t* bufferToFill, int64_t sampleOffset, int64_t samplesToRead) override;
+	virtual int64_t GetSampleCount() override;
 
 	virtual uint32_t GetChannelCount() override;
 	virtual uint32_t GetSampleRate() override;
