+++ Sandbox/src/Audio/MemoryAudioStream.h
@@ -12,8 +12,8 @@ public:
 	void LoadFromFile(const std::string& filePath);
 	void Dispose();
 
-	virtual size_t ReadSamples(int16_t* bufferToFill, size_t sampleOffset, size_t samplesToRead) override;
-	virtual size_t GetSampleCount() override;
+	virtual int64_t ReadSamples(int16_t* bufferToFill, int64_t sampleOffset, int64_t samplesToRead) override;
+	virtual int64_t GetSampleCount() override;
 
 	virtual uint32_t GetChannelCount() override;
 	virtual uint32_t GetSampleRate() override;
@@ -27,7 +27,7 @@ private:
 	uint32_t channelCount;
 	uint32_t sampleRate;
 
-	size_t sampleCount = 0;
+	uint64_t sampleCount = 0;
 	int16_t* sampleData = nullptr;
 
 	uint32_t GetFileMagic(const wchar_t* filePath);
